chainlit/requirements.txt
# Chatbot de Medicamentos - OpenAI Version
# Dependências necessárias

# Chainlit - Framework para interface de chatbot
chainlit>=1.0.0

# OpenAI - Conexão com a API da OpenAI
openai>=1.0.0

# Gerenciamento de variáveis de ambiente
python-dotenv>=1.0.0

# HTTP requests para API de medicamentos
requests>=2.28.0

# Dependências para análise de dados (opcionais, para futuras expansões)
pandas>=1.5.0
numpy>=1.24.0

# Outras dependências úteis
typing-extensions>=4.0.0